requirements.txt
annotated-types==0.6.0
anyio==4.6.2
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
distro==1.9.0
fastapi==0.112.2
filelock==3.16.1
fsspec==2024.10.0
h11==0.14.0
httpcore==1.0.2
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
jiter==0.6.1
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai==1.52.1
openai-whisper==20240930
pip==24.2
pydantic==2.8.2
pydantic_core==2.20.1
PySocks==1.7.1
python-dotenv==0.21.0
python-multipart==0.0.9
regex==2024.9.11
requests==2.32.3
setuptools==75.1.0
sniffio==1.3.0
starlette==0.38.2
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.66.5
typing_extensions==4.11.0
urllib3==2.2.3
uvicorn==0.20.0
wheel==0.44.0
